refactor!: make PolicySettings a struct
Prior to this commit, `PolicySettings` was just a type alias for a
`serde_json::Map<String, serde_json::Value>`.

Now this is a struct that still wraps the serde map.
However, by being a real type, we can implement different `try_from`
traits that hugely simplify the code consuming this type.
